Update main.py
Fixed some issues with double data points at beginning or end of dataset. Just chop off the first and last index of each dataset. 

when running the gamma analysis you will have to close out of each window that pops up in order to continue the pdf generation. Then when completed you will have to go to whatever folder you are using for this program in order to find the pdf file. This fixes one of the major topics required in the main readme.